Minor GCC-10 static analyzer fixes
GCC 10 (pre-release) has a static analyzer that can detect some common
memory errors.  It's not always quite sane, yet, but it did detect a
couple spots where we could strengthen our code.

Add in additional NULL checks as appropriate, and a single header which
is no longer included automatically by the new libstdc++ in GCC10+.

All changes are harmless to existing GCC4 or 9.